Separates channel-safely plugin functions and core logic to diff files
