+++ plugins/CMakeLists.txt
@@ -100,7 +100,7 @@ if(BUILD_SUPPORTED)
     dfhack_plugin(changeitem changeitem.cpp)
     dfhack_plugin(changelayer changelayer.cpp)
     dfhack_plugin(changevein changevein.cpp)
-    dfhack_plugin(channel-safely channel-safely.h channel-safely.cpp)
+    dfhack_plugin(channel-safely channel-safely.h channel-safely.cpp channel-safely-plugin.cpp)
     dfhack_plugin(cleanconst cleanconst.cpp)
     dfhack_plugin(cleaners cleaners.cpp)
     dfhack_plugin(cleanowned cleanowned.cpp)
